logica de excluir de todos os cruds
excluir de todos os cruds e ajuste da navegação

+++ LojaDeLivros/src/main/java/com/spring/ifpb/repository/UsuarioRepository.java
@@ -6,10 +6,11 @@ import org.springframework.stereotype.Repository;
 import com.spring.ifpb.model.Cliente;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Cliente, Integer>{
+public interface UsuarioRepository extends JpaRepository<Cliente, Long>{
 
 	Cliente findByCpf(Long cpf);
 	Cliente findByEmail(String email);
+	
 	
 
 }
